initial branch from cte-parse.  Now decompose source into hierarchical structure
